1. Created (abstract)activities. 2. Created CRUD example. 3. Connected with H2 database. 3. Service and Controller created again from the beginning 4. Service and Controller for user created

+++ src/main/java/com/projectbaby/Project/baby/repository/BabyRepository.java
@@ -4,6 +4,6 @@ import com.projectbaby.Project.baby.model.Baby;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface BabyRepository extends JpaRepository<Baby, Integer> {
+public interface BabyRepository extends JpaRepository<Baby, Long> {
 
 }
